fix: fixed colors and types
